Replace explict `cast_to_<ring-name>` by dynamic_cast
Note that the explicit virtual methods do exactly that: they return nullptr
at the Ring level, and the in <ring-name> overload to `this`. That behavior
is a special case of C++'s `dynamic_cast` operation.

There is technically a slight performance hit incurred when downcasting, as
opposed to having the named virtual methods. However, it is ephymeral; and
in the context of garbage collected objects, and given that everything *is*
virtual, the few casts would not introduce a slow down.

On the other hand, they greatly clen up the code.